Disallow Void operations, tuples, function call arguments + tests + FunCall return changes
Change how FunCall unifies with the expected type. Long story short, a while ago, FunCall would unify the exp_type with the output of the function call. Works well with func() + 1 and x = func(), but doesn't work well when its a function call as a statement in a function, as that will change the function return type. This was fixed via **kwargs and a return_funcall keyword parameter, but this introduced other bugs. Turns out, we almost always want FunCall to unify with exp_type, except when it's just a pure statement where the return of the function does not matter. So, now we don't pass a correct exp_type down from StmtNode to FunCallNode to avoid the function call from modifying the return type.